:Debuggging

+++ app/entities/responses/AuthResponse.scala
@@ -3,4 +3,4 @@ package app.entities.responses
 import org.joda.time.DateTime
 
 
-case class AuthResponse(id:Integer,username:String,token:String,dateCreated:DateTime)
+case class AuthResponse(id:Long,username:String,token:String,dateCreated:DateTime)
